test: move slow tests to one place
