+++ src/decoder.ts
@@ -612,6 +612,4 @@ export const __tests = {
   parseRegisters,
   exceptions,
   fixWindowsPath,
-  buildTarch,
-  buildTarget,
 } as const;
